src/bot/requirements.txt
Django==2.0.2
djangorestframework==3.9.1
mysql-connector-python==8.0.21
numpy==1.19.2
opencv-contrib-python==4.4.0.46
pytesseract==0.3.6
requests==2.22.0
pymysql==0.10.1
python-dotenv==0.14.0
PyNaCl==1.4.0
six==1.15.0
cffi==1.14.3
pycparser==2.20
matplotlib==3.3.2
aiogram==2.10.1
Pygments==2.7.2
uWSGI==2.0.19.1

